Create db.models.announcements
