+++ api/routes/announcements.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Session
 from sqlalchemy import desc, select
 from db.database import get_db
 from utils.oauth2 import get_current_user
-from db.models.users import Announcement
+from db.models.announcements import Announcement
 from api.api_models.announcements import AnnouncementCreate, AnnouncementResponse, PaginatedResponse, AnnouncementUpdate
 from utils.permissions import is_admin
