+++ db/models/announcements.py
@@ -1,0 +1,20 @@
+from db.database import Base
+from sqlalchemy import Column, String, Integer, TIMESTAMP, text
+from sqlalchemy.orm import relationship
+from sqlalchemy import ForeignKey
+
+
+
+class Announcement(Base):
+    __tablename__ = 'announcements'
+    id = Column(Integer, primary_key=True, nullable=False)
+    title = Column(String, nullable=False)
+    content = Column(String, nullable=False)
+    image_url = Column(String)
+    created_at = Column(
+        TIMESTAMP(timezone=True), nullable=False, server_default=text("now()"))
+    updated_at = Column(TIMESTAMP(timezone=True),
+                        nullable=False, server_default=text('now()'))
+    user_id = Column(
+        Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
+    user = relationship("User")
